create some 'Unity Build' cpp files, to work-around setuptools on Windows creating >32k command-line see 'won't fix' in https://bugs.python.org/issue1539295
