+++ MANIFEST.in
@@ -5,9 +5,11 @@ recursive-include Extras *.h
 recursive-include Extras *.hpp
 recursive-include src *.h
 recursive-include src *.hpp
+recursive-include src *.cpp
 recursive-include examples/pybullet/gym *.*
 include examples/ThirdPartyLibs/enet/unix.c
 include examples/OpenGLWindow/*.*
 recursive-include examples/SharedMemory/plugins *.*
 recursive-include examples/ThirdPartyLibs/glad *.*
 include examples/ThirdPartyLibs/enet/win32.c
+recursive-include examples/ThirdPartyLibs/Eigen *
