+++ setup.py
@@ -72,6 +72,9 @@ else:
 	  include_dirs += NP_DIRS
 
 sources = ["examples/pybullet/pybullet.c"]\
++["src/btLinearMathAll.cpp"]\
++["src/btBulletCollisionAll.cpp"]\
++["src/btBulletDynamicsAll.cpp"]\
 +["examples/ExampleBrowser/InProcessExampleBrowser.cpp"]\
 +["examples/TinyRenderer/geometry.cpp"]\
 +["examples/TinyRenderer/model.cpp"]\
@@ -180,156 +183,6 @@ sources = ["examples/pybullet/pybullet.c"]\
 +["examples/ExampleBrowser/GwenGUISupport/gwenUserInterface.cpp"]\
 +["examples/ExampleBrowser/GwenGUISupport/GwenParameterInterface.cpp"]\
 +["examples/ExampleBrowser/GwenGUISupport/GwenTextureWindow.cpp"]\
-+["src/LinearMath/btAlignedAllocator.cpp"]\
-+["src/LinearMath/btGeometryUtil.cpp"]\
-+["src/LinearMath/btSerializer.cpp"]\
-+["src/LinearMath/btVector3.cpp"]\
-+["src/LinearMath/btConvexHull.cpp"]\
-+["src/LinearMath/btPolarDecomposition.cpp"]\
-+["src/LinearMath/btSerializer64.cpp"]\
-+["src/LinearMath/btConvexHullComputer.cpp"]\
-+["src/LinearMath/btQuickprof.cpp"]\
-+["src/LinearMath/btThreads.cpp"]\
-+["src/LinearMath/TaskScheduler/btTaskScheduler.cpp"]\
-+["src/LinearMath/TaskScheduler/btThreadSupportPosix.cpp"]\
-+["src/LinearMath/TaskScheduler/btThreadSupportWin32.cpp"]\
-+["src/BulletCollision/BroadphaseCollision/btAxisSweep3.cpp"]\
-+["src/BulletCollision/BroadphaseCollision/btDbvt.cpp"]\
-+["src/BulletCollision/BroadphaseCollision/btOverlappingPairCache.cpp"]\
-+["src/BulletCollision/BroadphaseCollision/btBroadphaseProxy.cpp"]\
-+["src/BulletCollision/BroadphaseCollision/btDbvtBroadphase.cpp"]\
-+["src/BulletCollision/BroadphaseCollision/btQuantizedBvh.cpp"]\
-+["src/BulletCollision/BroadphaseCollision/btCollisionAlgorithm.cpp"]\
-+["src/BulletCollision/BroadphaseCollision/btDispatcher.cpp"]\
-+["src/BulletCollision/BroadphaseCollision/btSimpleBroadphase.cpp"]\
-+["src/BulletCollision/CollisionDispatch/SphereTriangleDetector.cpp"]\
-+["src/BulletCollision/CollisionDispatch/btCompoundCollisionAlgorithm.cpp"]\
-+["src/BulletCollision/CollisionDispatch/btHashedSimplePairCache.cpp"]\
-+["src/BulletCollision/CollisionDispatch/btActivatingCollisionAlgorithm.cpp"]\
-+["src/BulletCollision/CollisionDispatch/btCompoundCompoundCollisionAlgorithm.cpp"]\
-+["src/BulletCollision/CollisionDispatch/btInternalEdgeUtility.cpp"]\
-+["src/BulletCollision/CollisionDispatch/btBox2dBox2dCollisionAlgorithm.cpp"]\
-+["src/BulletCollision/CollisionDispatch/btConvex2dConvex2dAlgorithm.cpp"]\
-+["src/BulletCollision/CollisionDispatch/btManifoldResult.cpp"]\
-+["src/BulletCollision/CollisionDispatch/btBoxBoxCollisionAlgorithm.cpp"]\
-+["src/BulletCollision/CollisionDispatch/btConvexConcaveCollisionAlgorithm.cpp"]\
-+["src/BulletCollision/CollisionDispatch/btSimulationIslandManager.cpp"]\
-+["src/BulletCollision/CollisionDispatch/btBoxBoxDetector.cpp"]\
-+["src/BulletCollision/CollisionDispatch/btConvexConvexAlgorithm.cpp"]\
-+["src/BulletCollision/CollisionDispatch/btSphereBoxCollisionAlgorithm.cpp"]\
-+["src/BulletCollision/CollisionDispatch/btCollisionDispatcher.cpp"]\
-+["src/BulletCollision/CollisionDispatch/btConvexPlaneCollisionAlgorithm.cpp"]\
-+["src/BulletCollision/CollisionDispatch/btSphereSphereCollisionAlgorithm.cpp"]\
-+["src/BulletCollision/CollisionDispatch/btCollisionObject.cpp"]\
-+["src/BulletCollision/CollisionDispatch/btDefaultCollisionConfiguration.cpp"]\
-+["src/BulletCollision/CollisionDispatch/btSphereTriangleCollisionAlgorithm.cpp"]\
-+["src/BulletCollision/CollisionDispatch/btCollisionWorld.cpp"]\
-+["src/BulletCollision/CollisionDispatch/btEmptyCollisionAlgorithm.cpp"]\
-+["src/BulletCollision/CollisionDispatch/btUnionFind.cpp"]\
-+["src/BulletCollision/CollisionDispatch/btCollisionWorldImporter.cpp"]\
-+["src/BulletCollision/CollisionDispatch/btGhostObject.cpp"]\
-+["src/BulletCollision/NarrowPhaseCollision/btContinuousConvexCollision.cpp"]\
-+["src/BulletCollision/NarrowPhaseCollision/btGjkEpaPenetrationDepthSolver.cpp"]\
-+["src/BulletCollision/NarrowPhaseCollision/btPolyhedralContactClipping.cpp"]\
-+["src/BulletCollision/NarrowPhaseCollision/btConvexCast.cpp"]\
-+["src/BulletCollision/NarrowPhaseCollision/btGjkPairDetector.cpp"]\
-+["src/BulletCollision/NarrowPhaseCollision/btRaycastCallback.cpp"]\
-+["src/BulletCollision/NarrowPhaseCollision/btGjkConvexCast.cpp"]\
-+["src/BulletCollision/NarrowPhaseCollision/btMinkowskiPenetrationDepthSolver.cpp"]\
-+["src/BulletCollision/NarrowPhaseCollision/btSubSimplexConvexCast.cpp"]\
-+["src/BulletCollision/NarrowPhaseCollision/btGjkEpa2.cpp"]\
-+["src/BulletCollision/NarrowPhaseCollision/btPersistentManifold.cpp"]\
-+["src/BulletCollision/NarrowPhaseCollision/btVoronoiSimplexSolver.cpp"]\
-+["src/BulletCollision/CollisionShapes/btBox2dShape.cpp"]\
-+["src/BulletCollision/CollisionShapes/btConvexPolyhedron.cpp"]\
-+["src/BulletCollision/CollisionShapes/btShapeHull.cpp"]\
-+["src/BulletCollision/CollisionShapes/btBoxShape.cpp"]\
-+["src/BulletCollision/CollisionShapes/btConvexShape.cpp"]\
-+["src/BulletCollision/CollisionShapes/btSphereShape.cpp"]\
-+["src/BulletCollision/CollisionShapes/btBvhTriangleMeshShape.cpp"]\
-+["src/BulletCollision/CollisionShapes/btConvexTriangleMeshShape.cpp"]\
-+["src/BulletCollision/CollisionShapes/btStaticPlaneShape.cpp"]\
-+["src/BulletCollision/CollisionShapes/btCapsuleShape.cpp"]\
-+["src/BulletCollision/CollisionShapes/btCylinderShape.cpp"]\
-+["src/BulletCollision/CollisionShapes/btStridingMeshInterface.cpp"]\
-+["src/BulletCollision/CollisionShapes/btCollisionShape.cpp"]\
-+["src/BulletCollision/CollisionShapes/btEmptyShape.cpp"]\
-+["src/BulletCollision/CollisionShapes/btTetrahedronShape.cpp"]\
-+["src/BulletCollision/CollisionShapes/btCompoundShape.cpp"]\
-+["src/BulletCollision/CollisionShapes/btHeightfieldTerrainShape.cpp"]\
-+["src/BulletCollision/CollisionShapes/btTriangleBuffer.cpp"]\
-+["src/BulletCollision/CollisionShapes/btConcaveShape.cpp"]\
-+["src/BulletCollision/CollisionShapes/btMinkowskiSumShape.cpp"]\
-+["src/BulletCollision/CollisionShapes/btTriangleCallback.cpp"]\
-+["src/BulletCollision/CollisionShapes/btConeShape.cpp"]\
-+["src/BulletCollision/CollisionShapes/btMultiSphereShape.cpp"]\
-+["src/BulletCollision/CollisionShapes/btTriangleIndexVertexArray.cpp"]\
-+["src/BulletCollision/CollisionShapes/btConvex2dShape.cpp"]\
-+["src/BulletCollision/CollisionShapes/btMultimaterialTriangleMeshShape.cpp"]\
-+["src/BulletCollision/CollisionShapes/btTriangleIndexVertexMaterialArray.cpp"]\
-+["src/BulletCollision/CollisionShapes/btConvexHullShape.cpp"]\
-+["src/BulletCollision/CollisionShapes/btOptimizedBvh.cpp"]\
-+["src/BulletCollision/CollisionShapes/btTriangleMesh.cpp"]\
-+["src/BulletCollision/CollisionShapes/btConvexInternalShape.cpp"]\
-+["src/BulletCollision/CollisionShapes/btPolyhedralConvexShape.cpp"]\
-+["src/BulletCollision/CollisionShapes/btTriangleMeshShape.cpp"]\
-+["src/BulletCollision/CollisionShapes/btConvexPointCloudShape.cpp"]\
-+["src/BulletCollision/CollisionShapes/btScaledBvhTriangleMeshShape.cpp"]\
-+["src/BulletCollision/CollisionShapes/btSdfCollisionShape.cpp"]\
-+["src/BulletCollision/CollisionShapes/btMiniSDF.cpp"]\
-+["src/BulletCollision/CollisionShapes/btUniformScalingShape.cpp"]\
-+["src/BulletCollision/Gimpact/btContactProcessing.cpp"]\
-+["src/BulletCollision/Gimpact/btGImpactQuantizedBvh.cpp"]\
-+["src/BulletCollision/Gimpact/btTriangleShapeEx.cpp"]\
-+["src/BulletCollision/Gimpact/gim_memory.cpp"]\
-+["src/BulletCollision/Gimpact/btGImpactBvh.cpp"]\
-+["src/BulletCollision/Gimpact/btGImpactShape.cpp"]\
-+["src/BulletCollision/Gimpact/gim_box_set.cpp"]\
-+["src/BulletCollision/Gimpact/gim_tri_collision.cpp"]\
-+["src/BulletCollision/Gimpact/btGImpactCollisionAlgorithm.cpp"]\
-+["src/BulletCollision/Gimpact/btGenericPoolAllocator.cpp"]\
-+["src/BulletCollision/Gimpact/gim_contact.cpp"]\
-+["src/BulletDynamics/Dynamics/btDiscreteDynamicsWorld.cpp"]\
-+["src/BulletDynamics/Dynamics/btRigidBody.cpp"]\
-+["src/BulletDynamics/Dynamics/btSimulationIslandManagerMt.cpp"]\
-+["src/BulletDynamics/Dynamics/btDiscreteDynamicsWorldMt.cpp"]\
-+["src/BulletDynamics/Dynamics/btSimpleDynamicsWorld.cpp"]\
-+["src/BulletDynamics/ConstraintSolver/btBatchedConstraints.cpp"]\
-+["src/BulletDynamics/ConstraintSolver/btConeTwistConstraint.cpp"]\
-+["src/BulletDynamics/ConstraintSolver/btGeneric6DofSpringConstraint.cpp"]\
-+["src/BulletDynamics/ConstraintSolver/btSliderConstraint.cpp"]\
-+["src/BulletDynamics/ConstraintSolver/btContactConstraint.cpp"]\
-+["src/BulletDynamics/ConstraintSolver/btHinge2Constraint.cpp"]\
-+["src/BulletDynamics/ConstraintSolver/btSolve2LinearConstraint.cpp"]\
-+["src/BulletDynamics/ConstraintSolver/btFixedConstraint.cpp"]\
-+["src/BulletDynamics/ConstraintSolver/btHingeConstraint.cpp"]\
-+["src/BulletDynamics/ConstraintSolver/btTypedConstraint.cpp"]\
-+["src/BulletDynamics/ConstraintSolver/btGearConstraint.cpp"]\
-+["src/BulletDynamics/ConstraintSolver/btNNCGConstraintSolver.cpp"]\
-+["src/BulletDynamics/ConstraintSolver/btUniversalConstraint.cpp"]\
-+["src/BulletDynamics/ConstraintSolver/btGeneric6DofConstraint.cpp"]\
-+["src/BulletDynamics/ConstraintSolver/btPoint2PointConstraint.cpp"]\
-+["src/BulletDynamics/ConstraintSolver/btGeneric6DofSpring2Constraint.cpp"]\
-+["src/BulletDynamics/ConstraintSolver/btSequentialImpulseConstraintSolver.cpp"]\
-+["src/BulletDynamics/ConstraintSolver/btSequentialImpulseConstraintSolverMt.cpp"]\
-+["src/BulletDynamics/MLCPSolvers/btDantzigLCP.cpp"]\
-+["src/BulletDynamics/MLCPSolvers/btLemkeAlgorithm.cpp"]\
-+["src/BulletDynamics/MLCPSolvers/btMLCPSolver.cpp"]\
-+["src/BulletDynamics/Featherstone/btMultiBody.cpp"]\
-+["src/BulletDynamics/Featherstone/btMultiBodyDynamicsWorld.cpp"]\
-+["src/BulletDynamics/Featherstone/btMultiBodyJointMotor.cpp"]\
-+["src/BulletDynamics/Featherstone/btMultiBodyGearConstraint.cpp"]\
-+["src/BulletDynamics/Featherstone/btMultiBodyConstraint.cpp"]\
-+["src/BulletDynamics/Featherstone/btMultiBodyFixedConstraint.cpp"]\
-+["src/BulletDynamics/Featherstone/btMultiBodyPoint2Point.cpp"]\
-+["src/BulletDynamics/Featherstone/btMultiBodyConstraintSolver.cpp"]\
-+["src/BulletDynamics/Featherstone/btMultiBodyMLCPConstraintSolver.cpp"]\
-+["src/BulletDynamics/Featherstone/btMultiBodyJointLimitConstraint.cpp"]\
-+["src/BulletDynamics/Featherstone/btMultiBodySliderConstraint.cpp"]\
-+["src/BulletDynamics/Featherstone/btMultiBodySphericalJointMotor.cpp"]\
-+["src/BulletDynamics/Vehicle/btRaycastVehicle.cpp"]\
-+["src/BulletDynamics/Vehicle/btWheelInfo.cpp"]\
-+["src/BulletDynamics/Character/btKinematicCharacterController.cpp"]\
 +["src/Bullet3Common/b3AlignedAllocator.cpp"]\
 +["src/Bullet3Common/b3Logging.cpp"]\
 +["src/Bullet3Common/b3Vector3.cpp"]\
