+++ src/BulletCollision/Gimpact/btGImpactBvhStructs.h
@@ -28,28 +28,7 @@ subject to the following restrictions:
 
 #include "btBoxCollision.h"
 #include "btTriangleShapeEx.h"
-
-//! Overlapping pair
-struct GIM_PAIR
-{
-	int m_index1;
-	int m_index2;
-	GIM_PAIR()
-	{
-	}
-
-	GIM_PAIR(const GIM_PAIR& p)
-	{
-		m_index1 = p.m_index1;
-		m_index2 = p.m_index2;
-	}
-
-	GIM_PAIR(int index1, int index2)
-	{
-		m_index1 = index1;
-		m_index2 = index2;
-	}
-};
+#include "gim_pair.h" //for GIM_PAIR
 
 ///GIM_BVH_DATA is an internal GIMPACT collision structure to contain axis aligned bounding box
 struct GIM_BVH_DATA
